Split test app into per-screen modules
Made-with: Cursor
