+++ e2e-tests/MaestroTestApp/src/test_cases_screen.ts
@@ -1,0 +1,12 @@
+import { TEST_CASES } from './test_cases';
+
+export function showTestCases() {
+  const app = document.getElementById('app')!;
+  app.innerHTML = '<h1>Test Cases</h1>';
+  Object.values(TEST_CASES).forEach((tc) => {
+    const btn = document.createElement('button');
+    btn.textContent = tc.title;
+    btn.addEventListener('click', tc.show);
+    app.appendChild(btn);
+  });
+}
